backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
